+++ src/range_proof.rs
@@ -1,2 +1,8 @@
 pub mod reciprocal;
 pub mod u64_proof;
+
+#[cfg(feature = "serde")]
+mod serializable;
+
+#[cfg(feature = "serde")]
+pub use serializable::SerializableProof;
