+++ src/wnla.rs
@@ -1,10 +1,12 @@
 ///! Definition and implementation of the Bulletproofs++ weight norm linear argument protocol.
 use std::ops::{Add, Mul};
-use k256::{AffinePoint, ProjectivePoint, Scalar};
+use k256::{ProjectivePoint, Scalar};
 use merlin::Transcript;
-use serde::{Deserialize, Serialize};
 use crate::transcript;
 use crate::util::*;
+
+#[cfg(feature = "serde")]
+mod serializable;
 
 /// Represents public information to be used in weight norm linear argument protocol.
 #[derive(Clone, Debug)]
@@ -26,37 +28,6 @@ pub struct Proof {
     pub x: Vec<ProjectivePoint>,
     pub l: Vec<Scalar>,
     pub n: Vec<Scalar>,
-}
-
-/// Represent serializable version of  weight norm linear argument proof (uses AffinePoint instead of ProjectivePoint).
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct SerializableProof {
-    pub r: Vec<AffinePoint>,
-    pub x: Vec<AffinePoint>,
-    pub l: Vec<Scalar>,
-    pub n: Vec<Scalar>,
-}
-
-impl From<&SerializableProof> for Proof {
-    fn from(value: &SerializableProof) -> Self {
-        return Proof {
-            r: value.r.iter().map(|r_val| ProjectivePoint::from(r_val)).collect::<Vec<ProjectivePoint>>(),
-            x: value.x.iter().map(|x_val| ProjectivePoint::from(x_val)).collect::<Vec<ProjectivePoint>>(),
-            l: value.l.clone(),
-            n: value.n.clone(),
-        };
-    }
-}
-
-impl From<&Proof> for SerializableProof {
-    fn from(value: &Proof) -> Self {
-        return SerializableProof {
-            r: value.r.iter().map(|r_val| r_val.to_affine()).collect::<Vec<AffinePoint>>(),
-            x: value.x.iter().map(|x_val| x_val.to_affine()).collect::<Vec<AffinePoint>>(),
-            l: value.l.clone(),
-            n: value.n.clone(),
-        };
-    }
 }
 
 impl WeightNormLinearArgument {
